+++ src/main/kotlin/com/github/blarc/ai/commits/intellij/plugin/AICommitAction.kt
@@ -2,9 +2,10 @@ package com.github.blarc.ai.commits.intellij.plugin
 
 import com.github.blarc.ai.commits.intellij.plugin.notifications.Notification
 import com.github.blarc.ai.commits.intellij.plugin.notifications.sendNotification
-import com.github.blarc.ai.commits.intellij.plugin.settings.AppSettings2
+import com.github.blarc.ai.commits.intellij.plugin.settings.ProjectSettings
 import com.intellij.openapi.actionSystem.AnAction
 import com.intellij.openapi.actionSystem.AnActionEvent
+import com.intellij.openapi.components.service
 import com.intellij.openapi.project.DumbAware
 import com.intellij.openapi.vcs.VcsDataKeys
 import com.intellij.openapi.vcs.ui.CommitMessage
@@ -12,7 +13,9 @@ import com.intellij.vcs.commit.AbstractCommitWorkflowHandler
 
 class AICommitAction : AnAction(), DumbAware {
     override fun actionPerformed(e: AnActionEvent) {
-        val llmClient = AppSettings2.instance.getActiveLLMClientConfiguration()
+        val project = e.project ?: return
+
+        val llmClient = project.service<ProjectSettings>().getActiveLLMClientConfiguration()
         if (llmClient == null) {
             sendNotification(Notification.clientNotSet())
             return
@@ -30,7 +33,6 @@ class AICommitAction : AnAction(), DumbAware {
             return
         }
 
-        val project = e.project ?: return
         llmClient.generateCommitMessage(commitWorkflowHandler, commitMessage, project)
     }
 }