+++ src/main/kotlin/com/github/blarc/ai/commits/intellij/plugin/settings/AppSettings2.kt
@@ -68,7 +68,7 @@ class AppSettings2 : PersistentStateComponent<AppSettings2> {
 
     @XMap
     var prompts = DefaultPrompts.toPromptsMap()
-    var activePrompt = prompts["basic"]!!
+    var activePrompt = DefaultPrompts.BASIC.prompt
 
     var appExclusions: Set<String> = setOf()